[sys] refactored os
