+++ src/vm/os/os_v1.c
@@ -4,7 +4,7 @@
  * Created:
  *   11 December 2023 at 9:06:14 GMT+3
  * Modified:
- *   February 9, 2025 at 11:12:55 AM GMT+3
+ *   February 10, 2025 at 5:22:14 PM GMT+3
  *
  */
 /*
@@ -32,9 +32,9 @@
 
 #include "sys/memory/memory_v1.h"
 
-#include "vm/v1/pointer/pointer_v1.h"
-#include "vm/v1/types/string/string_v1.h"
-#include "vm/v1/virtual/virtual_v1.h"
+#include "vm/pointer/pointer_v1.h"
+#include "vm/types/string/string_v1.h"
+#include "vm/virtual/virtual_v1.h"
 
 #include <stdio.h>
 #include <stdlib.h>
