+++ src/vm/types/stack/stack_v1.c
@@ -4,7 +4,7 @@
  * Created:
  *   11 December 2023 at 9:06:14 GMT+3
  * Modified:
- *   February 9, 2025 at 11:07:50 AM GMT+3
+ *   February 10, 2025 at 5:22:14 PM GMT+3
  *
  */
 /*
@@ -30,8 +30,8 @@
 
 #include "sys/list/list_v1.h"
 
-#include "vm/v1/pointer/pointer_v1.h"
-#include "vm/v1/virtual/virtual_v1.h"
+#include "vm/pointer/pointer_v1.h"
+#include "vm/virtual/virtual_v1.h"
 
 #define DEFAULT_SIZE 0x100
