+++ src/vm/types/string_pointer/string_pointer_v1.h
