+++ src/vm/virtual/virtual_v1.h
