+++ tests/vm/test.h
