+++ tests/vm/test_list.c
