+++ tests/vm/test_vm_v1.c
@@ -4,7 +4,7 @@
  * Created:
  *   11 December 2023 at 9:06:14 GMT+3
  * Modified:
- *   February 9, 2025 at 1:15:59 PM GMT+3
+ *   February 10, 2025 at 5:22:14 PM GMT+3
  *
  */
 /*
@@ -27,11 +27,11 @@
 #include "test_vm_v1.h"
 
 #include "sys/options/options_v1.h"
-#include "vm/v1/pointer/pointer_v1.h"
-#include "vm/v1/types/data/data_v1.h"
-#include "vm/v1/types/stack/stack_v1.h"
-#include "vm/v1/virtual/virtual_v1.h"
-#include "vm/v1/vm_v1.h"
+#include "vm/pointer/pointer_v1.h"
+#include "vm/types/data/data_v1.h"
+#include "vm/types/stack/stack_v1.h"
+#include "vm/virtual/virtual_v1.h"
+#include "vm/vm_v1.h"
 
 #define DEFAULT_SIZE 0x100
